requirements.txt
agno
pywebview
openai
groq
anthropic
google-genai
pydantic
python-dotenv
requests
pyinstaller
sqlalchemy
pypdf
lancedb
tantivy
fastembed
